Enable db logging for user controller sql executions
