+++ .idea/workspace.xml
@@ -1,10 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="53d0fc5e-109e-4ba3-9217-18bcb6d3d79a" name="Changes" comment="Update workspace dictionary">
+    <list default="true" id="53d0fc5e-109e-4ba3-9217-18bcb6d3d79a" name="Changes" comment="Enable db logging for session controller sql executions">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/models.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/tests/conftest.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/conftest.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/tests/controllers/conftest.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/controllers/conftest.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/tests/controllers/test_users.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/controllers/test_users.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/tests/models/conftest.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/models/conftest.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/tests/support/declarations.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/support/declarations.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/tests/support/mocks/clients.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/tests/support/mocks/clients.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/utils/sql.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/utils/sql.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -150,13 +154,6 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1673428541001</updated>
-    </task>
-    <task id="LOCAL-00042" summary="Update routing engine">
-      <created>1676886210862</created>
-      <option name="number" value="00042" />
-      <option name="presentableId" value="LOCAL-00042" />
-      <option name="project" value="LOCAL" />
-      <updated>1676886210862</updated>
     </task>
     <task id="LOCAL-00043" summary="Allow pycharm workspace settings exportation">
       <created>1676941688363</created>
@@ -494,7 +491,14 @@
       <option name="project" value="LOCAL" />
       <updated>1678838191486</updated>
     </task>
-    <option name="localTasksCounter" value="91" />
+    <task id="LOCAL-00091" summary="Enable db logging for session controller sql executions">
+      <created>1679007090975</created>
+      <option name="number" value="00091" />
+      <option name="presentableId" value="LOCAL-00091" />
+      <option name="project" value="LOCAL" />
+      <updated>1679007090976</updated>
+    </task>
+    <option name="localTasksCounter" value="92" />
     <servers />
   </component>
   <component name="UnknownFeatures">
@@ -512,7 +516,6 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="Update conftests" />
     <MESSAGE value="Update settings.py" />
     <MESSAGE value="Update routes.py" />
     <MESSAGE value="Update validators.py &amp; commons.py" />
@@ -537,6 +540,7 @@
     <MESSAGE value="Update test support to count for :remove_message routines" />
     <MESSAGE value="Add test unit for :remove_message routines" />
     <MESSAGE value="Update workspace dictionary" />
-    <option name="LAST_COMMIT_MESSAGE" value="Update workspace dictionary" />
+    <MESSAGE value="Enable db logging for session controller sql executions" />
+    <option name="LAST_COMMIT_MESSAGE" value="Enable db logging for session controller sql executions" />
   </component>
 </project>
