+++ app/tests/conftest.py
@@ -3,7 +3,6 @@ import platform
 import asyncio as io
 import pathlib as plib
 
-from .support.models import create_user
 from ..utils.db_connect import db_connector, drop_schema
 from ..utils.commons import get_attribute_values
 from ..settings import DB_CONFIGS, APP_NAME
@@ -54,17 +53,3 @@ def purge_db():
     yield
     for cls in reversed(tuple(get_attribute_values(models, exceptions=('_Model', 'Activity')))):
         cls.delete().execute()
-
-@pytest.fixture(scope='class')
-def user(request):
-    _user = create_user()
-    request.cls.user = _user
-    yield _user
-    _user.delete_instance()
-
-@pytest.fixture(scope='class')
-def other_user(request):
-    _user = create_user(email='other_user@email')
-    request.cls.other_user = _user
-    yield _user
-    _user.delete_instance()
