+++ app/tests/models/conftest.py
@@ -1,6 +1,20 @@
 import pytest
 
-from ..support.models import create_channel, create_message, create_activity
+from ..support.models import create_channel, create_message, create_activity, create_user
+
+@pytest.fixture(scope='class')
+def user(request):
+    _user = create_user()
+    request.cls.user = _user
+    yield _user
+    _user.delete_instance()
+
+@pytest.fixture(scope='class')
+def other_user(request):
+    _user = create_user(email='other_user@email')
+    request.cls.other_user = _user
+    yield _user
+    _user.delete_instance()
 
 @pytest.fixture(scope='class')
 def channel(request, user):
